Tempus: Split StepperFactory.hpp.
Split StepperFactory.hpp into StepperFactory_decl.hpp and
StepperFactory_impl.hpp fro ETI.

 * Made createMultiSteppers() name uniform with, e.g.,
   changed to createStepper().
