+++ packages/tempus/src/Tempus_IntegratorBasic_impl.hpp
@@ -103,7 +103,7 @@ void IntegratorBasic<Scalar>::setStepper(
     std::string stepperName = integratorPL_->get<std::string>("Stepper Name");
 
     RCP<ParameterList> stepperPL = Teuchos::sublist(tempusPL_,stepperName,true);
-    stepper_ = sf->createMultiSteppers(stepperPL, models);
+    stepper_ = sf->createStepper(stepperPL, models);
   } else {
     stepper_->createSubSteppers(models);
   }
@@ -413,7 +413,7 @@ void IntegratorBasic<Scalar>::startIntegrator()
     integratorStatus_ = Status::FAILED;
     return;
   }
-  
+
   //set the Abs/Rel tolerance
   auto cs = solutionHistory_->getCurrentState();
   cs->setTolRel(timeStepControl_->getMaxRelError());
