+++ packages/tempus/src/Tempus_StepperRKBase.hpp
@@ -67,10 +67,10 @@ protected:
 
   // For Embedded RK
   bool useEmbedded_;
-  Teuchos::RCP<Thyra::VectorBase<Scalar> >               ee_;
-  Teuchos::RCP<Thyra::VectorBase<Scalar> >               abs_u0;
-  Teuchos::RCP<Thyra::VectorBase<Scalar> >               abs_u;
-  Teuchos::RCP<Thyra::VectorBase<Scalar> >               sc;
+  Teuchos::RCP<Thyra::VectorBase<Scalar> >  ee_;
+  Teuchos::RCP<Thyra::VectorBase<Scalar> >  abs_u0;
+  Teuchos::RCP<Thyra::VectorBase<Scalar> >  abs_u;
+  Teuchos::RCP<Thyra::VectorBase<Scalar> >  sc;
 
   /// The current Runge-Kutta stage number, {0,...,s-1}.  -1 indicates outside stage loop.
   int stageNumber_;
